requirements.txt
ultralytics
torch
torchvision
pillow
faiss-cpu
numpy
matplotlib